Refactored sockserver into its own header.
